pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fhfa-data-manager"
version = "0.1.0"
description = "Tools for managing public data from FHFA."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "ipykernel>=6.30.1",
    "pandas>=2.3.2",
    "polars>=1.33.1",
    "pyarrow>=21.0.0",
    "pyodbc>=5.2.0",
    "python-decouple>=3.8",
    "requests>=2.32.5",
    "pymupdf>=1.23.0",
]

[tool.setuptools.packages.find]
where = ["src"]
